reorganize project structure by updating include paths and renaming source files
