--- include/common.h
@@ -1,8 +1,0 @@
-//
-// Created by Kazz on 22/06/2025.
-//
-
-#ifndef COMMON_H
-#define COMMON_H
-
-#endif //COMMON_H
