--- include/interpreter.h
@@ -1,8 +1,0 @@
-//
-// Created by Kazz on 22/06/2025.
-//
-
-#ifndef INTERPRETER_H
-#define INTERPRETER_H
-
-#endif //INTERPRETER_H
