--- src/interpreter.c
@@ -1,3 +1,0 @@
-//
-// Created by Kazz on 22/06/2025.
-//
