+++ src/lexer/lexer.c
@@ -1,4 +1,4 @@
-#include "../include/lexer.h"
+#include "lexer.h"
 #include <stdlib.h>
 #include <string.h>
 #include <ctype.h>
