+++ src/parser/parser.c
@@ -1,4 +1,4 @@
-#include "../include/parser.h"
+#include "parser.h"
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
